refactor(network): Move network config from main to network mod
