+++ src/main.rs
@@ -1,15 +1,11 @@
 use bevy::prelude::*;
-use bevy_ggrs::prelude::*;
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
 use rand::Rng as _;
 
-use bevy_matchbox::prelude::*;
-
 mod core;
 mod entities;
+mod network;
 mod utils;
-
-type Config = bevy_ggrs::GgrsConfig<u8, PeerId>;
 
 fn main() {
     let mut app = App::new();
@@ -30,8 +26,7 @@ fn main() {
     )
     .add_plugins(core::CorePlugins)
     .add_plugins(entities::EntitiesPlugins)
-    .add_systems(Startup, start_matchbox_socket)
-    .add_systems(Update, wait_for_players);
+    .add_plugins(network::NetworkPlugin);
 
     if cfg!(debug_assertions) {
         app.add_plugins(WorldInspectorPlugin::new());
@@ -40,51 +35,8 @@ fn main() {
     app.run();
 }
 
-fn init(mut worldgen_events: EventWriter<core::worldgen::GenerateWorldEvent>) {
+fn _init(mut worldgen_events: EventWriter<core::worldgen::GenerateWorldEvent>) {
     worldgen_events.send(core::worldgen::GenerateWorldEvent {
         seed: rand::thread_rng().gen(),
     });
-}
-
-fn start_matchbox_socket(mut commands: Commands) {
-    let room_url = "ws://127.0.0.1:3536/extreme_bevy?next=2";
-    info!("connecting to matchbox server: {room_url}");
-    commands.insert_resource(MatchboxSocket::new_unreliable(room_url));
-}
-
-fn wait_for_players(mut commands: Commands, mut socket: ResMut<MatchboxSocket>) {
-    if socket.get_channel(0).is_err() {
-        return; // we've already started
-    }
-
-    // Check for new connections
-    socket.update_peers();
-    let players = socket.players();
-
-    let num_players = 2;
-    if players.len() < num_players {
-        return; // wait for more players
-    }
-
-    info!("All peers have joined, going in-game");
-
-    let mut session_builder = SessionBuilder::<Config>::new()
-        .with_num_players(num_players)
-        .with_input_delay(2);
-
-    for (i, player) in players.into_iter().enumerate() {
-        session_builder = session_builder
-            .add_player(player, i)
-            .expect("failed to add player");
-    }
-
-    // move the channel out of the socket (required because GGRS takes ownership of it)
-    let channel = socket.take_channel(0).unwrap();
-
-    // start the GGRS session
-    let ggrs_session = session_builder
-        .start_p2p_session(channel)
-        .expect("failed to start session");
-
-    commands.insert_resource(bevy_ggrs::Session::P2P(ggrs_session));
 }
